# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/item_variants:
    get:
      tags:
        - ItemVariants
      summary: List ItemVariants
      description: Get a paginated list of ItemVariants with optional filtering and sorting
      operationId: listItemVariants
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemVariantList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ItemVariants
      summary: Create ItemVariant
      description: Create a new ItemVariant
      operationId: createItemVariant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemVariantRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemVariant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/uom_conversions:
    get:
      tags:
        - UomConversions
      summary: List UomConversions
      description: Get a paginated list of UomConversions with optional filtering and sorting
      operationId: listUomConversions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UomConversionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - UomConversions
      summary: Create UomConversion
      description: Create a new UomConversion
      operationId: createUomConversion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUomConversionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UomConversion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ItemVariant:
      type: object
      required:
        - id
        - company_id
        - item_id
        - sku
        - variant_label
        - options
        - is_default
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        sku:
          type: string
          maxLength: 60
        variant_label:
          type: string
          maxLength: 200
        options:
          type: object
        barcode:
          nullable: true
          type: string
          maxLength: 60
        is_default:
          type: boolean
        weight_per_unit:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateItemVariantRequest:
      type: object
      required:
        - company_id
        - item_id
        - sku
        - variant_label
        - options
        - is_default
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        sku:
          type: string
          maxLength: 60
        variant_label:
          type: string
          maxLength: 200
        options:
          type: object
        barcode:
          nullable: true
          type: string
          maxLength: 60
        is_default:
          type: boolean
        weight_per_unit:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    ItemVariantList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemVariant'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    UomConversion:
      type: object
      required:
        - id
        - company_id
        - from_uom_id
        - to_uom_id
        - factor
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        from_uom_id:
          type: string
          format: uuid
        to_uom_id:
          type: string
          format: uuid
        factor:
          type: string
        metadata:
          type: object
...
    CreateUomConversionRequest:
      type: object
      required:
        - company_id
        - from_uom_id
        - to_uom_id
        - factor
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        from_uom_id:
          type: string
          format: uuid
        to_uom_id:
          type: string
          format: uuid
        factor:
          type: string
        metadata:
          type: object
...
    UomConversionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UomConversion'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []