# src/presentation/http/item_variant_handler.rs
//! ItemVariant REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{ItemVariantService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateItemVariantDto, UpdateItemVariantDto, PatchItemVariantDto, ItemVariantResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum ItemVariantError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for ItemVariantError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for ItemVariantError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "ITEMVARIANT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "ITEMVARIANT_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "ITEMVARIANT_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "ITEMVARIANT_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for ItemVariant.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /item_variants | List with pagination |
/// | POST | /item_variants | Create new |
/// | GET | /item_variants/:id | Get by ID |
/// | PUT | /item_variants/:id | Full update |
/// | PATCH | /item_variants/:id | Partial update |
/// | DELETE | /item_variants/:id | Soft delete |
/// | POST | /item_variants/bulk | Bulk create |
/// | POST | /item_variants/upsert | Upsert |
/// | GET | /item_variants/trash | List deleted |
/// | POST | /item_variants/:id/restore | Restore |
/// | DELETE | /item_variants/empty | Empty trash |
/// | GET | /item_variants/:id/deleted | Get deleted by ID |
/// | DELETE | /item_variants/trash/:id | Permanent delete from trash |
/// | GET | /item_variants/count | Count active entities |
/// | GET | /item_variants/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(ItemVariantService::with_repository(repository));
/// let router = create_item_variant_routes(service);
/// ```
pub fn create_item_variant_routes(service: Arc<ItemVariantService>) -> Router {
    BackboneCrudHandler::<ItemVariantService, ItemVariant, CreateItemVariantDto, UpdateItemVariantDto, ItemVariantResponseDto>::routes(
        service,
        "/item_variants",
    )
}

/// Create Axum router with only the read (GET) endpoints for ItemVariant.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_item_variant_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_item_variant_read_routes(service: Arc<ItemVariantService>) -> Router {
    BackboneCrudHandler::<ItemVariantService, ItemVariant, CreateItemVariantDto, UpdateItemVariantDto, ItemVariantResponseDto>::read_routes(
        service,
        "/item_variants",
    )
}

/// Create Axum router with only the write (mutation) endpoints for ItemVariant.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_item_variant_write_routes(service: Arc<ItemVariantService>) -> Router {
    BackboneCrudHandler::<ItemVariantService, ItemVariant, CreateItemVariantDto, UpdateItemVariantDto, ItemVariantResponseDto>::write_routes(
        service,
        "/item_variants",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_item_variant_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<ItemVariantService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_item_variant_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}