# src/presentation/http/uom_conversion_handler.rs
//! UomConversion REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{UomConversionService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateUomConversionDto, UpdateUomConversionDto, PatchUomConversionDto, UomConversionResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum UomConversionError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for UomConversionError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for UomConversionError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "UOMCONVERSION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "UOMCONVERSION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "UOMCONVERSION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "UOMCONVERSION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for UomConversion.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /uom_conversions | List with pagination |
/// | POST | /uom_conversions | Create new |
/// | GET | /uom_conversions/:id | Get by ID |
/// | PUT | /uom_conversions/:id | Full update |
/// | PATCH | /uom_conversions/:id | Partial update |
/// | DELETE | /uom_conversions/:id | Soft delete |
/// | POST | /uom_conversions/bulk | Bulk create |
/// | POST | /uom_conversions/upsert | Upsert |
/// | GET | /uom_conversions/trash | List deleted |
/// | POST | /uom_conversions/:id/restore | Restore |
/// | DELETE | /uom_conversions/empty | Empty trash |
/// | GET | /uom_conversions/:id/deleted | Get deleted by ID |
/// | DELETE | /uom_conversions/trash/:id | Permanent delete from trash |
/// | GET | /uom_conversions/count | Count active entities |
/// | GET | /uom_conversions/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(UomConversionService::with_repository(repository));
/// let router = create_uom_conversion_routes(service);
/// ```
pub fn create_uom_conversion_routes(service: Arc<UomConversionService>) -> Router {
    BackboneCrudHandler::<UomConversionService, UomConversion, CreateUomConversionDto, UpdateUomConversionDto, UomConversionResponseDto>::routes(
        service,
        "/uom_conversions",
    )
}

/// Create Axum router with only the read (GET) endpoints for UomConversion.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_uom_conversion_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_uom_conversion_read_routes(service: Arc<UomConversionService>) -> Router {
    BackboneCrudHandler::<UomConversionService, UomConversion, CreateUomConversionDto, UpdateUomConversionDto, UomConversionResponseDto>::read_routes(
        service,
        "/uom_conversions",
    )
}

/// Create Axum router with only the write (mutation) endpoints for UomConversion.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_uom_conversion_write_routes(service: Arc<UomConversionService>) -> Router {
    BackboneCrudHandler::<UomConversionService, UomConversion, CreateUomConversionDto, UpdateUomConversionDto, UomConversionResponseDto>::write_routes(
        service,
        "/uom_conversions",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_uom_conversion_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<UomConversionService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_uom_conversion_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}